Moved surprise me color generation to the theme factory instead of the downloadPageViewModel, so each factory can pick it's own surprise theme colors now.
Also added an iOS6, metro, and flat theme factory.

The factories now load in a JSON file which serves as the basis for modifications. The theme can then be modified with the array of colors passed in to the factory. This means that there is a lot less code, as you can just modify a few colours, rather than rebuilding the entire theme.